+++ sdk/storage/azure-storage-blobs/test/ut/block_blob_client_test.cpp
@@ -154,7 +154,7 @@ namespace Azure { namespace Storage { namespace Test {
 
       Azure::Storage::Blobs::ListBlobsOptions options;
       options.Prefix = m_blobName;
-      for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasMorePages();
+      for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasPage();
            pageResult.MoveToNextPage())
       {
         for (const auto& blob : pageResult.Blobs)
@@ -322,7 +322,7 @@ namespace Azure { namespace Storage { namespace Test {
     Azure::Storage::Blobs::ListBlobsOptions options;
     options.Prefix = blobName;
     options.Include = Blobs::Models::ListBlobsIncludeFlags::Versions;
-    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasMorePages();
+    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasPage();
          pageResult.MoveToNextPage())
     {
       for (const auto& blob : pageResult.Blobs)
@@ -976,7 +976,7 @@ namespace Azure { namespace Storage { namespace Test {
 
     Azure::Storage::Blobs::ListBlobsOptions options;
     options.Prefix = blobName;
-    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasMorePages();
+    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasPage();
          pageResult.MoveToNextPage())
     {
       for (const auto& blob : pageResult.Blobs)
@@ -1002,7 +1002,7 @@ namespace Azure { namespace Storage { namespace Test {
     EXPECT_FALSE(properties.IsAccessTierInferred.Value());
     EXPECT_TRUE(properties.AccessTierChangedOn.HasValue());
 
-    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasMorePages();
+    for (auto pageResult = m_blobContainerClient->ListBlobs(options); pageResult.HasPage();
          pageResult.MoveToNextPage())
     {
       for (const auto& blob : pageResult.Blobs)